+++ src/python/esgcet/esgunpublish.py
@@ -1,4 +1,5 @@
-import esgcet.unpublish as upub
+from esgcet.unpublish_globus import ESGUnpublishGlobus
+from esgcet.unpublish_solr import ESGUnpublishSolr
 import os
 import sys
 import json
@@ -23,7 +24,8 @@ def get_args():
     parser.add_argument("--data-node", dest="data_node", default=None, help="Specify data node.")
     parser.add_argument("--certificate", "-c", dest="cert", default=None,
                         help="Use the following certificate file in .pem form for unpublishing (use a myproxy login to generate).")
-    parser.add_argument("--delete", dest="delete", action="store_true", help="Specify deletion of dataset (default is retraction).")
+    parser.add_argument("--delete",  action="store_true", help="Specify deletion of dataset (default is retraction).")
+    parser.add_argument("--deprecate",  action="store_true", help="Specify deprecation of dataset (default is retraction).")
     parser.add_argument("--dset-id", dest="dset_id", default=None,
                         help="Dataset ID for dataset to be retracted or deleted.")
     parser.add_argument("--map", dest="map", default=None, nargs="+",
@@ -150,13 +152,23 @@ def run():
         verbose = True
         silent = False
 
-    args = { "delete": d, 
-             "data_node": data_node, 
-             "index_node": index_node, 
-             "cert": cert, 
-             "auth" :auth, 
+    args = { "delete": d,
+             "data_node": data_node,
+             "index_node": index_node,
+             "cert": cert,
+             "auth" :auth,
              "verbose" : verbose,
-             "silent" :silent }
+             "silent" :silent,
+             }
+
+    if config.get("globus_index", False):
+        args["index_UUID"] = config.get("index_UUID", "")
+        upub = ESGUnpublishGlobus()
+    else:
+        upub = ESGUnpublishSolr()
+
+    args["dry_run"] = config.get("dry_run", False)
+    args["deprecate"] = a.deprecate
 
     if len(dset_id) > 0:
         args["dataset_id_lst"] = [dset_id]