+++ src/python/esgcet/generic_pub.py
@@ -1,6 +1,8 @@
 from esgcet.mapfile import ESGPubMapConv
 from esgcet.mkd_non_nc import ESGPubMKDNonNC
-from esgcet.update import ESGPubUpdate
+from esgcet.update_solr import ESGUpdateSolr
+from esgcet.update_globus import ESGUpdateGlobus
+
 from esgcet.index_pub import ESGPubIndex
 import sys
 import esgcet.logger as logger
@@ -28,6 +30,7 @@ class BasePublisher(object):
         self.verify = argdict["verify"]
         self.mountpoints = argdict["mountpoints"]
         self.project = argdict["proj"]
+        self.dry_run = argdict.get("dry_run", False)
         self.publog = log.return_logger('Generic Non-NetCDF Publisher', self.silent, self.verbose)
 
     def cleanup(self):
@@ -59,7 +62,11 @@ class BasePublisher(object):
         return out_json_data
 
     def update(self, json_data):
-        up = ESGPubUpdate(self.index_node, self.cert, silent=self.silent, verbose=self.verbose, verify=self.verify, auth=self.auth)
+    
+        if self.argdict.get("globus_index", False):
+            up = ESGUpdateGlobus(self.argdict.get("index_UUID"), json_data[0]["data_node"], silent=self.silent, verbose=self.verbose, dry_run=self.dry_run)
+        else:
+            up = ESGUpdateSolr(self.index_node, silent=self.silent, verbose=self.verbose, verify=self.verify)
         try:
             up.run(json_data)
         except Exception as ex:
@@ -72,13 +79,25 @@ class BasePublisher(object):
         if self.argdict["enable_archive"]:
             arch_cfg = { "length" : int(self.argdict["archive_path_length"]) , 
                           "archive_path" : self.argdict["archive_path"]}
+        print(f"VERBOSE: {self.verbose}")
+        # TODO: support solr and Globus using the globus_index argument
 
-        ip = ESGPubIndex(self.index_node, self.cert, silent=self.silent, verbose=self.verbose, verify=self.verify, auth=self.auth, arch_cfg=arch_cfg)
+        globuspub = self.argdict.get("globus_index", False)
+        
+        if globuspub:
+            index_node = ""
+        else:
+            index_node = dataset_records[0]["index_node"]
+        ip = ESGPubIndex(index_node=index_node, UUID=self.argdict["index_UUID"],  silent=self.silent, verbose=self.verbose, verify=self.verify, auth=self.auth, arch_cfg=arch_cfg, dry_run=self.dry_run)
+            
         rc = True
         try:
-            rc = ip.do_publish(dataset_records)
+            if globuspub:
+                rc = ip.do_globus(dataset_records)
+            else:
+                rc = ip.do_publish(dataset_records)
         except Exception as ex:
-            self.publog.exception("Failed to publish to index node")
+            self.publog.exception("Failed to publish to index.")
             self.cleanup()
             exit(1)
         return rc